Update NVidia cub to 1.13.0
ref:cb80c51c7bd64e6e176b0f7743e60a6478090c75

+++ contrib/libs/nvidia/cub/cub/warp/specializations/warp_reduce_shfl.cuh
@@ -1,7 +1,7 @@
 /******************************************************************************
  * Copyright (c) 2011, Duane Merrill.  All rights reserved.
  * Copyright (c) 2011-2018, NVIDIA CORPORATION.  All rights reserved.
- * 
+ *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
  *     * Redistributions of source code must retain the above copyright
@@ -12,7 +12,7 @@
  *     * Neither the name of the NVIDIA CORPORATION nor the
  *       names of its contributors may be used to endorse or promote products
  *       derived from this software without specific prior written permission.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
  * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -33,11 +33,12 @@
 
 #pragma once
 
+#include "../../config.cuh"
 #include "../../thread/thread_operators.cuh"
 #include "../../util_ptx.cuh"
 #include "../../util_type.cuh"
-#include "../../util_macro.cuh"
-#include "../../util_namespace.cuh"
+
+#include <stdint.h>
 
 /// Optional outer namespace(s)
 CUB_NS_PREFIX
@@ -96,13 +97,13 @@ struct WarpReduceShfl
     //---------------------------------------------------------------------
 
     /// Lane index in logical warp
-    unsigned int lane_id;
+    int lane_id;
 
     /// Logical warp index in 32-thread physical warp
-    unsigned int warp_id;
+    int warp_id;
 
     /// 32-thread physical warp member mask of logical warp
-    unsigned int member_mask;
+    uint32_t member_mask;
 
 
     //---------------------------------------------------------------------
@@ -113,7 +114,7 @@ struct WarpReduceShfl
     __device__ __forceinline__ WarpReduceShfl(
         TempStorage &/*temp_storage*/)
     {
-        lane_id = LaneId();
+        lane_id = static_cast<int>(LaneId());
         warp_id = 0;
         member_mask = 0xffffffffu >> (CUB_WARP_THREADS(PTX_ARCH) - LOGICAL_WARP_THREADS);
 
@@ -347,13 +348,13 @@ struct WarpReduceShfl
         KeyValuePair<KeyT, ValueT> output;
 
         KeyT other_key = ShuffleDown<LOGICAL_WARP_THREADS>(input.key, offset, last_lane, member_mask);
-        
+
         output.key = input.key;
         output.value = ReduceStep(
-            input.value, 
-            cub::Sum(), 
-            last_lane, 
-            offset, 
+            input.value,
+            cub::Sum(),
+            last_lane,
+            offset,
             Int2Type<IsInteger<ValueT>::IS_SMALL_UNSIGNED>());
 
         if (input.key != other_key)
